Trinity.C: Network: refactoring. posix handshake wip.
