+++ src/Trinity.C/src/Network/Client/TrinityClient.h
@@ -10,6 +10,12 @@
 #include <wchar.h>
 #include <thread>
 #include <TrinityCommon.h>
+#include <Trinity/Diagnostics/Log.h>
+
+#include "Network/Network.h"
+
+#include "Trinity/Hash/NonCryptographicHash.h"
+#include "Network/ProtocolConstants.h"
 
 namespace Trinity
 {
@@ -65,5 +71,34 @@ namespace Trinity
             }
             return 0;
         }
+
+        inline bool ClientSocketHandshake(uint64_t clientsocket)
+        {
+            char* send_buff      = (char*)malloc(HANDSHAKE_MESSAGE_LENGTH + sizeof(int32_t));
+            *(int32_t*)send_buff = HANDSHAKE_MESSAGE_LENGTH;
+            memcpy(send_buff + sizeof(int32_t), HANDSHAKE_MESSAGE_CONTENT, HANDSHAKE_MESSAGE_LENGTH);
+
+            if (!ClientSend(clientsocket, send_buff, sizeof(int32_t) + HANDSHAKE_MESSAGE_LENGTH))
+            {
+                Diagnostics::WriteLine(Diagnostics::LogLevel::Error, "ClientSocket: Server rejected handshake request message. Thread Id = {0}, Socket = {1}", std::this_thread::get_id(), clientsocket);
+                goto handshake_check_fail;
+            }
+
+            if (WaitForAckPackage(clientsocket) != TrinityErrorCode::E_SUCCESS)
+            {
+                Diagnostics::WriteLine(Diagnostics::LogLevel::Error, "ClientSocket: Incorrect server handshake ACK message. Thread Id = {0}, Socket = {1}", std::this_thread::get_id(), clientsocket);
+                goto handshake_check_fail;
+            }
+
+            //handshake_check_success: cleanup and return now.
+            free(send_buff);
+            return true;
+
+        handshake_check_fail:
+            free(send_buff);
+            closesocket((SOCKET)clientsocket);
+            return false;
+        }
+
     }
 }
