+++ src/Trinity.C/src/Network/Client/posix/ClientSocket.h
