+++ src/Trinity.C/src/Network/Client/windows/ClientSocket.cpp
@@ -8,8 +8,6 @@
 #include "Network/Server/iocp/TrinitySocketServer.h"
 #include "Network/SocketOptionsHelper.h"
 #include <thread>
-#include "Trinity/Hash/NonCryptographicHash.h"
-#include "Network/ProtocolConstants.h"
 
 namespace Trinity
 {
@@ -30,34 +28,6 @@ namespace Trinity
 
             // The client side already have heartbeat packages. Don't enable keepalive messages.
             return SetSocketOptions(clientsocket, /*enable_keepalive:*/ false, /*disable_sendbuf:*/false);
-        }
-
-        bool ClientSocketHandshake(uint64_t clientsocket)
-        {
-            char* send_buff      = (char*)malloc(HANDSHAKE_MESSAGE_LENGTH + sizeof(int32_t));
-            *(int32_t*)send_buff = HANDSHAKE_MESSAGE_LENGTH;
-            memcpy(send_buff + sizeof(int32_t), HANDSHAKE_MESSAGE_CONTENT, HANDSHAKE_MESSAGE_LENGTH);
-
-            if (!ClientSend(clientsocket, send_buff, sizeof(int32_t) + HANDSHAKE_MESSAGE_LENGTH))
-            {
-                Diagnostics::WriteLine(Diagnostics::LogLevel::Error, "ClientSocket: Server rejected handshake request message. Thread Id = {0}, Socket = {1}", std::this_thread::get_id(), clientsocket);
-                goto handshake_check_fail;
-            }
-
-            if (WaitForAckPackage(clientsocket) != TrinityErrorCode::E_SUCCESS)
-            {
-                Diagnostics::WriteLine(Diagnostics::LogLevel::Error, "ClientSocket: Incorrect server handshake ACK message. Thread Id = {0}, Socket = {1}", std::this_thread::get_id(), clientsocket);
-                goto handshake_check_fail;
-            }
-
-            //handshake_check_success: cleanup and return now.
-            free(send_buff);
-            return true;
-
-        handshake_check_fail:
-            free(send_buff);
-            closesocket((SOCKET)clientsocket);
-            return false;
         }
 
         bool ClientSocketConnect(uint64_t clientsocket, uint32_t ip, uint16_t port)
