--- src/Trinity.C/src/Network/posix/common.cpp
@@ -1,16 +1,0 @@
-// Graph Engine
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
-//
-#include <os/os.h>
-#if !defined(TRINITY_PLATFORM_WINDOWS)
-#include <sys/socket.h>
-
-namespace Trinity
-{
-    namespace Network
-    {
-        
-    }
-}
-#endif
